fix(entity): add SMALLINT columnDefinition for OperationPlan.scale
Hibernate does not auto-map Kotlin Short to SMALLINT, causing
schema-validation failure on startup. Also fix Int→Short conversion
in OperationPlanService.createPlan().

+++ backend/game-app/src/main/kotlin/com/openlogh/service/OperationPlanService.kt
@@ -96,7 +96,7 @@ class OperationPlanService(
             targetStarSystemId = targetStarSystemId,
             status = OperationStatus.PENDING,
             participantFleetIds = participantFleetIds.toMutableList(),
-            scale = scale,
+            scale = scale.toShort(),
             issuedByOfficerId = issuedByOfficerId,
             // Phase 12 limitation: no tick accessor available at this layer.
             // issuedAtTick is always 0L. OPS-03 semantics do not depend on
